update variants, create useCustomVariant hook
