+++ packages/ui/hooks/useCustomVariant.ts
@@ -1,0 +1,7 @@
+import { type MantineTheme } from '@mantine/core'
+
+import { variantNames } from '~ui/theme/variants'
+
+export const useCustomVariant = () => {
+	return variantNames
+}
